+++ app/src/main/java/com/github/nakawai/newsreader/model/entity/Multimedia.kt
@@ -1,5 +1,6 @@
 package com.github.nakawai.newsreader.model.entity
 
+// TODO Add fields
 data class Multimedia(
     val url: String
 )
